Remove mockery from calendar service tests
