+++ test/services/calendar/index.test.js
@@ -1,32 +1,51 @@
 const assert = require('assert');
 const chai = require('chai');
+const appPromise = require('../../../src/app');
 
 const { expect } = chai;
-const mockery = require('mockery');
 
-const requestMock = require('./mock/mockResponses');
+const mockRequest = async (options) => {
+	const exampleResponse = {
+		links: {
+			self: 'https://dbildungscloud.de:3000/events',
+		},
+		data: [
+			{
+				type: 'event',
+				id: '0ef82d0a-0e20-465d-ae1e-850dc04d9432',
+				attributes: {
+					summary: 'tttttt',
+					location: 'Paul-Gerhardt-Gymnasium',
+					description: 'sdds',
+					dtstart: '2017-05-19T20:00:00.000Z',
+					dtend: '2017-05-19T20:00:00.000Z',
+					dtstamp: '2017-05-10T06:44:24.174Z',
+					uid: 'e40e1276-4a27-458c-8b64-ae549adadbc2',
+				},
+				relationships: {
+					'scope-ids': ['0000d231816abba584714c9e'],
+				},
+			},
+		],
+	};
+
+	return exampleResponse;
+};
 
 describe('calendar service', () => {
-	let app = null;
-	let calendarService = null;
+	let server;
+	let calendarService;
 
 	before(async () => {
-		mockery.enable({
-			warnOnReplace: false,
-			warnOnUnregistered: false,
-			useCleanCache: true,
-		});
-		mockery.registerMock('request-promise-native', requestMock);
 		// eslint-disable-next-line global-require
-		app = await require('../../../src/app')();
-		app.setup();
+		const app = await appPromise();
+		server = await app.listen(0);
 		calendarService = await app.service('calendar');
+		calendarService.send = async (options) => mockRequest(options);
 	});
 
-	after((done) => {
-		mockery.deregisterAll();
-		mockery.disable();
-		done();
+	after(async () => {
+		await server.close();
 	});
 
 	it('registered the calendar service', () => {
@@ -40,11 +59,11 @@ describe('calendar service', () => {
 		});
 
 		expect(result.length).to.be.above(0);
-		expect(result[0].title).to.not.be.undefined;
+		expect(result[0].title).to.not.equal(undefined);
 		expect(result[0].title).to.equal('tttttt');
-		expect(result[0].start).to.not.be.undefined;
+		expect(result[0].start).to.not.equal(undefined);
 		expect(result[0].start).to.equal(1495224000000);
-		expect(result[0].end).to.not.be.undefined;
+		expect(result[0].end).to.not.equal(undefined);
 		expect(result[0].end).to.equal(1495224000000);
 	});
 
@@ -59,6 +78,7 @@ describe('calendar service', () => {
 		};
 
 		const result = await calendarService.create(postBody, { payload: { userId: '0000d231816abba584714c9e' } });
+
 		expect(result.length).to.be.above(0);
 	});
 
@@ -81,6 +101,6 @@ describe('calendar service', () => {
 
 	it('DELETE /calendar', async () => {
 		const result = await calendarService.remove('exampleId', { payload: { userId: '0000d231816abba584714c9e' } });
-		expect(result).to.not.be.undefined;
+		expect(result).to.not.equal(undefined);
 	});
 });
